\name{dks}
\alias{dks}
\alias{eset}
\alias{pv.f}
\alias{pvr.f}
\docType{data}
\title{ A demonstration data set for package dualKS. }
\description{
	This data set contains an \code{ExpressionSet} comprised of a 
	subset of data from the Gene Expression Omnibus (GEO) data set 
	GDS2126 (Rheumatoid arthritis: synovial tissues).  The data set 
	has been severely pruned down to allow quick execution of the 
	examples associated with this package.
}
\usage{data("dks")}
\format{
MIAME formatted data as an ExpressionSet.

}
\source{
	\code{http://www.ncbi.nlm.nih.gov/geo/gds/gds_browse.cgi?gds=2126}
}

\keyword{datasets}
